# Refactor AD2CP to improve speed and memory usage through removal of xr.merge