--- a/echopype/convert/parse_ad2cp.py
+++ b/echopype/convert/parse_ad2cp.py
@@ -722,37 +722,6 @@ class Ad2cpDataPacket:
                     self.parser.packets[-1].data[
                         "echosounder_raw_echogram"
                     ] = self.data["echosounder_index"]
-            elif field_name == "ahrs_rotation_matrix_m33":
-                self.data["ahrs_rotation_matrix"] = np.array(
-                    [
-                        self.data["ahrs_rotation_matrix_m11"],
-                        self.data["ahrs_rotation_matrix_m12"],
-                        self.data["ahrs_rotation_matrix_m13"],
-                        self.data["ahrs_rotation_matrix_m21"],
-                        self.data["ahrs_rotation_matrix_m22"],
-                        self.data["ahrs_rotation_matrix_m23"],
-                        self.data["ahrs_rotation_matrix_m31"],
-                        self.data["ahrs_rotation_matrix_m32"],
-                        self.data["ahrs_rotation_matrix_m33"],
-                    ]
-                )
-            elif field_name == "ahrs_quaternions_z":
-                self.data["ahrs_quaternions"] = np.array(
-                    [
-                        self.data["ahrs_quaternions_w"],
-                        self.data["ahrs_quaternions_x"],
-                        self.data["ahrs_quaternions_y"],
-                        self.data["ahrs_quaternions_z"],
-                    ]
-                )
-            elif field_name == "ahrs_gyro_z":
-                self.data["ahrs_gyro"] = np.array(
-                    [
-                        self.data["ahrs_gyro_x"],
-                        self.data["ahrs_gyro_y"],
-                        self.data["ahrs_gyro_z"],
-                    ]
-                )
         elif (
             self.data_record_format
             == HeaderOrDataRecordFormats.BOTTOM_TRACK_DATA_RECORD_FORMAT
@@ -1583,105 +1552,29 @@ class HeaderOrDataRecordFormats:
                 ],
             ),
             F(
-                "ahrs_rotation_matrix_m11",
+                "ahrs_rotation_matrix",
                 4,
                 FLOAT,
+                field_shape=[9],
+                field_dimensions=[Dimension.TIME, Dimension.MIJ],
                 field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
             ),
             F(
-                "ahrs_rotation_matrix_m12",
+                "ahrs_quaternions",
                 4,
                 FLOAT,
+                field_shape=[4],
+                field_dimensions=[Dimension.TIME, Dimension.WXYZ],
                 field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
             ),
             F(
-                "ahrs_rotation_matrix_m13",
+                "ahrs_gyro",
                 4,
                 FLOAT,
+                field_shape=[3],
+                field_dimensions=[Dimension.TIME, Dimension.XYZ],
                 field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
             ),
-            F(
-                "ahrs_rotation_matrix_m21",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_rotation_matrix_m22",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_rotation_matrix_m23",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_rotation_matrix_m31",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_rotation_matrix_m32",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_rotation_matrix_m33",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_quaternions_w",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_quaternions_x",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_quaternions_y",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_quaternions_z",
-                4,
-                FLOAT,
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_gyro_x",
-                4,
-                FLOAT,
-                field_units="degrees/s",
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_gyro_y",
-                4,
-                FLOAT,
-                field_units="degrees/s",
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            F(
-                "ahrs_gyro_z",
-                4,
-                FLOAT,
-                field_units="degrees/s",
-                field_exists_predicate=lambda packet: packet.data["ahrs_data_included"],
-            ),
-            # ("ahrs_gyro", 4, FLOAT, [3], lambda packet: packet.data["ahrs_data_included"]),
             F(
                 "percentage_good_data",
                 1,
@@ -1932,9 +1825,9 @@ class HeaderOrDataRecordFormats:
                 field_dimensions=[Dimension.TIME_ECHOSOUNDER_RAW, Dimension.SAMPLE],
                 field_exists_predicate=lambda packet: packet.is_echosounder_raw(),
             ),
-            # These next 2 fields are included so that the dimensions
-            # for this field can be determined
-            # based on the field name
+            # These next 2 fields are included so that the dimensions for these fields
+            #   can be determined based on the field name.
+            #   They are actually constructed in _postprocess.
             F(
                 "echosounder_raw_samples_i",
                 0,
@@ -1963,9 +1856,9 @@ class HeaderOrDataRecordFormats:
                 ],
                 field_exists_predicate=lambda packet: packet.is_echosounder_raw_transmit(),
             ),
-            # These next 2 fields are included so that the dimensions
-            # for this field can be determined
-            # based on the field name
+            # These next 2 fields are included so that the dimensions for these fields
+            #   can be determined based on the field name.
+            #   They are actually constructed in _postprocess.
             F(
                 "echosounder_raw_transmit_samples_i",
                 0,