# Summary

* Removes **all** xarray operations, including merges
    * Now uses numpy operations only
    * Removes xarray merge that would nullify the use of dask by loading everything into memory anyway
* Use numpy arrays everywhere
* Adds optimizations (see Benchmarks section)
* Generalizes/simplifies parser logic
    * Reorganizes packets in a way that makes saving chunks to disk easier
* Fixes type errors

# Benchmarks

Benchmarks of `echopype.open_raw` commit 8ad2e284bf910ff3bb78109a795e89e26c8e5cde with and without this PR applied

## Speed

Used Jupyter magic command `%%timeit -r 10 -n 1`

| File | File size | Mean time +/- std dev (before this PR) | Mean time +/- std dev (after this PR) | Speed (before this PR) | Speed (after this PR) |
| --- | --- | --- | --- | --- | --- |
| avg_bur_echo.366.00000.ad2cp | 866 KB | 27.9s +/- 741ms | 2.36s +/- 114ms | 31.04 KB/s | 366.95 KB/s |
| burst_echosoun.366.00000.ad2cp | 2400 KB | 50.8s +/- 718ms | 4.28s +/- 127ms | 47.24 KB/s | 560.75 KB/s |
| rawtest.076.00000.ad2cp | 1676 KB | 11s +/- 365ms | 1.38s +/- 73.3ms | 152.36 KB/s | 1214.49 KB/s |

## Memory usage

Used the [`memory_profiler`](https://pypi.org/project/memory-profiler/) package with command `mprof peak`

| File | File size | Memory usage (before this PR) | Memory usage (after this PR) |
| --- | --- | --- | --- |
| avg_bur_echo.366.00000.ad2cp | 866 KB | 382.754 MB | 171.719 MB |
| burst_echosoun.366.00000.ad2cp | 2400 KB | 837.312 MB | 206.074 MB |
| rawtest.076.00000.ad2cp | 1676 KB | 206.586 MB | 171.719 MB |
| rawEchosounderdata_0000.ad2cp | 1079940 KB | Did not complete, ran out of memory and crashed after 11210.910 MiB | 8166.355 MiB |

# Todo

- [x] Memory profiling