@emiliom 2021-12-16
The issue of memory expansion is discussed in #489