@leewujung 2022-06-03
Thanks @imranmaj, this is great! I tested the code with a few files and followed the flow of parsing and setting groups. Things look good and I think we can merge this now. I'll work on getting the variables into different `Beam_groupX` groups and discuss with you and @emiliom  once I have it!